requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.23
django-allauth==0.44.0
django-filter==2.4.0
django-import-export==2.5.0
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==2.10
ipython==7.24.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
numpy==1.20.3
oauthlib==3.1.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.24
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
